Corrections: an error in the 2D Gaussian kernel and something small in the total probability (it was not an error, but if we're looking for p(x), we shouldn't be using p(x) to calculate itself..)
Also, edited a bit the representation of the 2D Gaussian kernel, that doesn't use the UNDERBRACE, but keeps it more linear.
However, just above I have the commented the corrected version of the 2D Gaussian kernel, with the underbrace.
If you prefer the version WITH, then comment out the second version and uncomment the first.